FilesService: used BackgroundService
