+++ products/ASC.Files/Service/Thumbnail/Builder.cs
@@ -112,7 +112,7 @@ internal class Builder<T>
         }
         finally
         {
-            Launcher.Queue.TryRemove(fileData.FileId, out _);
+            FileDataQueue.Queue.TryRemove(fileData.FileId, out _);
         }
     }
